Setup delegates for the media picker.
