+++ WordPress/Classes/ViewRelated/Media/MediaLibraryPicker.swift
@@ -1,0 +1,23 @@
+import WPMediaPicker
+
+final class MediaLibraryPicker: NSObject {
+    private let dataSource = WPPHAssetDataSource()
+
+    weak var delegate: WPMediaPickerViewControllerDelegate?
+    private var blog: Blog?
+
+    func presentPicker(origin: UIViewController, blog: Blog) {
+        self.blog = blog
+        let options = WPMediaPickerOptions()
+        options.showMostRecentFirst = true
+        options.filter = [.all]
+        options.allowCaptureOfMedia = false
+
+
+        let picker = WPNavigationMediaPickerViewController(options: options)
+        picker.dataSource = dataSource
+        picker.delegate = delegate
+
+        origin.present(picker, animated: true, completion: nil)
+    }
+}
